requirements.txt
loguru==0.5.3
requests==2.22.0
webexteamssdk==1.2
webexteamsbot==0.1.4.2
python-dotenv==0.17.1
